Add state machine capabilities for use by modules
This change adds a new class called AnsibleModuleFSM which is built on
top of AnsibleModule via composition (not inheritance). The main advantage
of using this class is that a module developer can now write modules as
state machines which can be easier to manage especially for those that
have a complex decision tree.